Mainly edits to the load_msoa_locations (moved and edited for national use)
